dqwdqw

+++ src/main/java/com/ssafy/file/controller/BoardFileController.java
@@ -34,8 +34,8 @@ public class BoardFileController {
 
 	@PostMapping("/{id}")
 	public String uploadFile(@RequestBody MultipartFile file,
-		@PathVariable(value = "id") Long id) {
-		Long res = bFileService.save(id,file);
+		@PathVariable(value = "id") int id) {
+		int res = bFileService.save(id,file);
 		
 		return res != -1 ? "성공"  : "파일의 크기는 10MB를 넘을 수 없습니다";
 	}
